# api/openapi3_0.yaml
openapi: 3.0.0
info:
  title: Pandemos
  description: >-
    API for visualization of Infection Models
  version: "1"
security:
  - bearerAuth: []
servers:
  - url: 'https://api.example.com/api/v0'
  - url: 'http://localhost:8080/'
paths:
  /scenarios/:
    get:
      tags:
        - Simulations
      operationId: listScenarios
      responses:
        '200':
          description: list all Scenarios
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdList'
    post:
      tags:
        - Simulations
      operationId: createSimulations
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewScenario'
      responses:
        '200':
          description: create a new Scenario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ID'
  /scenarios/{scenario_id}/:
    get:
      tags:
        - Simulations
      operationId: getScenario
      parameters:
        - in: path
          name: scenario_id
          required: True
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: return the account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scenario'
    delete:
      tags:
        - Simulations
      operationId: deleteScenario
      parameters:
        - in: path
          name: scenario_id
          required: True
          schema:
            type: string
            format: uuid
      description: deletes the Scenario and all its runs
      responses:
        '202':
          description: scenario deleted
  /scenarios/{scenario_id}/simulations/:
    post:
      tags:
        - Simulations
      operationId: triggerSimulationRun
      parameters:
          - in: path
            name: scenario_id
            required: True
            schema:
              type: string
              format: uuid
      requestBody: 
        content: 
          application/zip:
            schema: 
              type: string
              format: binary
      responses:
        '201':
          description: return the account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ID'
  /scenarios/{scenario_id}/simulations/{runid}/:
    get:
      tags:
        - Simulations
      operationId: getSimulationRunStatus
      parameters:
        - in: path
          name: scenario_id
          required: True
          schema:
            type: string
            format: uuid
        - in: path
          name: runid
          required: True
          schema:
            type: string
            format: uuid
          
      responses:
        '200':
          description: return the simulation run status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimulationRunStatus'
    delete:
      tags:
        - Simulations
      operationId: deleteSimulationRun
      parameters:
        - in: path
          name: scenario_id
          required: True
          schema:
            type: string
            format: uuid
        - in: path
          name: runid
          required: True
          schema:
            type: string
            format: uuid
      description: deletes the run
      responses:
        '202':
          description: run deleted
  /scenarios/{scenario_id}/simulations/{runid}/infectiondata/:
    get:
      tags:
        - Simulations
      operationId: getInfectionData
      parameters:
        - in: path
          name: scenario_id
          required: True
          schema: 
            type: string
            format: uuid
        - in: path
          name: runid
          required: True
          schema: 
            type: string
            format: uuid
        - in: query
          required: false
          name: node
          schema:
            type: string
        - in: query
          required: false
          name: startDate
          schema:
            type: string
        - in: query
          required: false
          name: endDate
          schema:
            type: string
        - in: query
          required: false
          name: compartments
          description: Compartment Aggregation
          schema:
            type: string
        - in: query
          required: false
          name: aggregation_flag
          schema: 
            type: boolean
        - in: query
          required: false
          name: groups
          schema:
            type: array
            items: 
              type: string
      responses:
        '200':
          description: return the account
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Infectiondata'
  /scenarios/{scenario_id}/simulations/{runid}/gridcells/:
    get:
      tags:
        - Simulations
      operationId: listGridcells
      parameters:
        - in: path
          name: scenario_id
          required: True
          schema: 
            type: string
            format: uuid
        - in: path
          name: runid
          required: True
          schema: 
            type: string
            format: uuid
      responses:
        '200':
          description: return the account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdList'
  /scenarios/{scenario_id}/simulations/{runid}/movements/:
    get:
      tags:
        - Simulations
      operationId: listMovments
      parameters:
        - in: path
          name: scenario_id
          required: True
          schema: 
            type: string
            format: uuid
        - in: path
          name: runid
          required: True
          schema: 
            type: string
            format: uuid
      responses:
        '200':
          description: return the account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdList'
...
components:
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ID:
      type: object
      required:
        - id
      properties:
        id:
          type: string
          format: uuid
    Named:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
    IdList:
      type: array
      items: 
        type: string
        format: uuid
    NewScenario:
      allOf:
        - $ref: '#/components/schemas/Named'
        - type: object
          properties:
            modelId:
              type: string
              format: uuid
            modelParameters:
              type: array
              items: 
                 $ref: '#/components/schemas/ParameterValue'
            nodeListId:
              type: string
              format: uuid
            linkedInterventions:
              type: array
              items:
                 $ref: '#/components/schemas/ID'
    Scenario:
      allOf:
        - $ref: '#/components/schemas/ID'
        - $ref: '#/components/schemas/NewScenario'
...
    ParameterValueRange:
      type: object
      properties:
        groupId:
          type: string
        valueMinInclusiv:
          type: number
        valueMaxExclusiv:
          type: number
    ParameterValue:
      type: object
      required:
        - parameterId
      properties:
        parameterId:
          type: string
          format: uuid
        groups:
          type: array
          items:
             $ref: '#/components/schemas/ParameterValueRange'
        categories:
          type: array
          items:
             $ref: '#/components/schemas/ParameterValueRange'
...
    Infectiondata: 
      type: object 
      properties: 
        timestamp: 
          type: string
          format: timestamp
        node: 
          type: string
          format: uuid
        value: 
          type: number
          format: integer
          example: 1
...
    SimulationRunStatus:
      type: string
      enum: [TRIGGERED, COMPLETED]
...