# api/src/api/app/models/infectiondata.py
# coding: utf-8

from __future__ import annotations
from datetime import date, datetime  # noqa: F401

import re  # noqa: F401
from typing import Any, Dict, List, Optional  # noqa: F401

from pydantic import AnyUrl, BaseModel, EmailStr, Field, validator  # noqa: F401


class Infectiondata(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    Infectiondata - a model defined in OpenAPI

        timestamp: The timestamp of this Infectiondata [Optional].
        node: The node of this Infectiondata [Optional].
        value: The value of this Infectiondata [Optional].
    """

    timestamp: Optional[str] = Field(alias="timestamp", default=None)
    node: Optional[str] = Field(alias="node", default=None)
    value: Optional[float] = Field(alias="value", default=None)

Infectiondata.update_forward_refs()